pyproject.toml
[tool.poetry]
name = "secure-movie-django"
version = "0.1.0"
description = ""
authors = ["Emanuele Conforti <manu.conforti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
coreapi = "^2.3.3"
pyyaml = "^6.0.1"
mixer = "^7.2.2"
pytest-django = "^4.7.0"
coverage = "^7.3.2"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.58.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
